Extract the page for the editor settings to settings-editor component
